FY-190/refactor(shared): add areHookDepsEqual
